Add files via upload
It prevents computer power saving and is programmed for those who are troublesome to log in every time. If you press start, the shift key is pressed once a minute, if you press stop, it stops, and if you press quit, the program turns off. The language consisted of Hangul(Korean).